Moved components tests to component folders
